# benches: move sender to a spawned task in `watch` benchmark