## Motivation

The PR #5464 introduced a new bucket notifier to reduce contention in the `watch` channel. That change was motivated by the benchmark added in #5472. However, this benchmark does not present favorable results for the newer notifier if we compare it with the older solution.

Results with the new notifier:
`contention_resubscribe  time:   [81.507 ms 82.098 ms 82.740 ms]`
Results without it:
`contention_resubscribe  time:   [38.254 ms 38.938 ms 39.618 ms]`

This is most likely due to [sender task not running on a worker thread](https://github.com/tokio-rs/tokio/issues/5403#issuecomment-1432118096).

## Solution

Change the existing `watch` channel benchmark by moving the benchmarked loop to a spawned task, which will be executed by one of the worker threads. With this change the results are more comparable:

new notifier:
```
contention_resubscribe/10    time:   [930.74 µs 937.37 µs 942.98 µs]
contention_resubscribe/100   time:   [3.8498 ms 3.8565 ms 3.8634 ms]
contention_resubscribe/500   time:   [16.106 ms 16.111 ms 16.116 ms]
contention_resubscribe/1000  time:   [33.519 ms 33.881 ms 34.250 ms]
```
old notifier:
```
contention_resubscribe/10    time:   [901.93 µs 914.50 µs 925.82 µs]
contention_resubscribe/100   time:   [4.0990 ms 4.1044 ms 4.1106 ms]
contention_resubscribe/500   time:   [16.800 ms 16.810 ms 16.823 ms]
contention_resubscribe/1000  time:   [32.479 ms 32.497 ms 32.516 ms]
```

Note that since the addition of the new notifier, some optimizations (#5503) have been made to the underlying `Notify` primitive. Those also reduced lock contention so the new bucket solution no longer presents as significant improvements (or any improvement at all) in this benchmark as it might at the time of its addition. I still think it's worth keeping around as it improves performance with more worker threads. Here are the results using 24 worker threads (all of the above results were obtained using 6):

new notifier:
```
contention_resubscribe/10    time:   [881.98 µs 893.00 µs 902.81 µs]
contention_resubscribe/100   time:   [3.6447 ms 3.6536 ms 3.6621 ms]
contention_resubscribe/500   time:   [21.162 ms 21.317 ms 21.486 ms]
contention_resubscribe/1000  time:   [43.518 ms 43.927 ms 44.346 ms]
```
old notifier:
```
contention_resubscribe/10    time:   [930.67 µs 932.05 µs 933.49 µs]
contention_resubscribe/100   time:   [4.5549 ms 4.7225 ms 4.8852 ms]
contention_resubscribe/500   time:   [41.734 ms 41.789 ms 41.845 ms]
contention_resubscribe/1000  time:   [86.575 ms 86.634 ms 86.694 ms]
```
For benchmarking, I used a 12-core x86_64 machine running Linux 6.5.4.