@Darksonn 2023-09-28
After all, there's also a disadvantage to having it, namely that it makes the allocation a lot larger.