@satakuma 2023-10-01
I ran the original benchmark from the thread in #5403 ([tijsvd/tokio-watch-benchmark](https://github.com/tijsvd/tokio-watch-benchmark)) and the results don't seem to favor `BigNotify` either, although they vary quite a bit from run to run.

new notifier (`BigNotify`)
```
TOKIO WATCH
running single thread with spawned sender
start
avg=515.938µs  <490.603µs [ 492.752µs 495.533µs 545.552µs ] 582.422µs>
running multi thread with spawned sender
start
avg=485.695µs  <296.226µs [ 444.833µs 525.262µs 544.252µs ] 562.281µs>
CUSTOM WATCH
running multi thread with spawned sender
start
avg=236.832µs  <203.757µs [ 225.777µs 238.187µs 249.106µs ] 264.636µs>
```

old notifier (just `Notify`)
```
TOKIO WATCH
running single thread with spawned sender
start
avg=482.403µs  <461.733µs [ 462.903µs 464.243µs 509.432µs ] 513.813µs>
running multi thread with spawned sender
start
avg=383.811µs  <347.524µs [ 368.795µs 383.414µs 397.754µs ] 419.423µs>
CUSTOM WATCH
running multi thread with spawned sender
start
avg=218.534µs  <198.347µs [ 210.887µs 218.927µs 226.486µs ] 237.147µs>
``` 

The author of this benchmark mentioned that he used an 8-vcore machine, so I used 8 worker threads as well.

I think that if I can't observe any benefit from `BigNotify` using the benchmarking method which led to its addition *and* `BigNotify` makes the allocation larger, then it might be better to simply remove it. Alternatively, we can think about a heuristic restricting `BigNotify` to only runtimes with, for example, more than 12 worker threads, although I'm not sure if that's possible. Anyway, the custom implementation from the benchmark still runs much faster on my machine, so I think there is room for improvement when it comes to reducing contention.