@satakuma 2023-10-27
Sorry for the late response. I ran the benchmarks mentioned above on other machines with significantly weaker CPUs and `BigNotify` was actually more than 2x faster there, so I am no longer inclined to remove it. I guess its performance depends heavily on the hardware. Unfortunately, I don't have time right now to investigate it further. Let's leave it as is for now.